Improved MonsterDeath and more
-3Fixed Resourcebar rounding issue,
-Further implemented monsted death,
-Added a victory screen,
-Updated flow of the combat
-Main now initializes monsters as a mob (Monster party)
-Monster death sprite doesn't display backwards anymore.

+++ Character/ResourceBar.cpp
@@ -62,7 +62,7 @@ void ResourceBar::resize(const sf::Vector2f & newSize) {
 
 
 void ResourceBar::update() {
-	int unitsPerPixel = static_cast<int> (size.x / maxValue);
+	float unitsPerPixel = size.x / maxValue;
 	barInnerMeter.setSize(sf::Vector2f(static_cast<float>(currentValue * unitsPerPixel), size.y));
 }
